Fit a TabPFN model.

TabPFN() fits a model.

TabPFN(x, ...)

# Default S3 method
TabPFN(x, ...)

# S3 method for class 'data.frame'
TabPFN(x, y, ignore_pretraining_limits = FALSE, n_jobs = 1L, ...)

# S3 method for class 'matrix'
TabPFN(x, y, ignore_pretraining_limits = FALSE, n_jobs = 1L, ...)

# S3 method for class 'formula'
TabPFN(formula, data, ignore_pretraining_limits = FALSE, n_jobs = 1L, ...)

# S3 method for class 'recipe'
TabPFN(x, data, ignore_pretraining_limits = FALSE, n_jobs = 1L, ...)

Arguments

x

Depending on the context:

A data frame of predictors.

A matrix of predictors.

A recipe specifying a set of preprocessing steps created from recipes::recipe().

...

Not currently used, but required for extensibility.

y

When x is a data frame or matrix, y is the outcome specified as:

A data frame with 1 numeric column.

A matrix with 1 numeric column.

A numeric vector for regression or a factor for classification.

ignore_pretraining_limits

A logical. The maximum number of features 500 officially supported by the TabPFN python api. Set ignore_pretraining_limits to TRUE to override.

n_jobs

The number of parallel process workers.

formula

A formula specifying the outcome terms on the left-hand side, and the predictor terms on the right-hand side.

data

When a recipe or formula is used, data is specified as:

A data frame containing both the predictors and the outcome.

Value

A TabPFN object with elements:

fit: the python object containing the model.

levels: a character string of class levels (or NULL for regression)

training: a vector with the training set dimensions.

versions: a list of python and python package versions and information.

logging: any R or python messages produced by the computations.

blueprint: am object produced by hardhat::mold() used to process new data during prediction.

Details

Predictors do not require preprocessing; missing values and factor vectors are allowed.

Examples

predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- TabPFN(predictors, outcome)

# Formula interface
mod2 <- TabPFN(mpg ~ ., mtcars)

# Recipes interface
if (!rlang::is_installed("recipes")) {
 library(recipes)
 rec <-
  recipe(mpg ~ ., mtcars) |>
  step_log(disp)

 mod3 <- TabPFN(rec, mtcars)
 mod3
}