Predict using TabPFN

Predict using TabPFN

# S3 method for class 'TabPFN'
predict(object, new_data, ...)

# S3 method for class 'TabPFN'
augment(x, new_data, ...)

Arguments

object, x

A TabPFN object.

new_data

A data frame or matrix of new predictors.

...

Not used, but required for extensibility.

Value

predict() returns a tibble of predictions and augment() appends the columns in new_data. In either case, the number of rows in the tibble is guaranteed to be the same as the number of rows in new_data.

For regression data, the prediction is in the column .pred. For classification, the class predictions are in .pred_class and the probability estimates are in columns with the pattern .pred_{level} where level is the levels of the outcome factor vector.

Examples

car_train <- mtcars[ 1:20,   ]
car_test  <- mtcars[21:32, -1]

# Fit
mod <- TabPFN(mpg ~ cyl + log(drat), car_train)

# Predict, with preprocessing
predict(mod, car_test)
#> # A tibble: 12 × 1
#>    .pred
#>    <dbl>
#>  1  25.1
#>  2  12.2
#>  3  16.1
#>  4  16.2
#>  5  15.7
#>  6  29.4
#>  7  30.0
#>  8  24.9
#>  9  18.5
#> 10  21.0
#> 11  15.8
#> 12  30.0
augment(mod, car_test)
#> # A tibble: 12 × 11
#>    .pred   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  25.1     4 120.     97  3.7   2.46  20.0     1     0     3     1
#>  2  12.2     8 318     150  2.76  3.52  16.9     0     0     3     2
#>  3  16.1     8 304     150  3.15  3.44  17.3     0     0     3     2
#>  4  16.2     8 350     245  3.73  3.84  15.4     0     0     3     4
#>  5  15.7     8 400     175  3.08  3.84  17.0     0     0     3     2
#>  6  29.4     4  79      66  4.08  1.94  18.9     1     1     4     1
#>  7  30.0     4 120.     91  4.43  2.14  16.7     0     1     5     2
#>  8  24.9     4  95.1   113  3.77  1.51  16.9     1     1     5     2
#>  9  18.5     8 351     264  4.22  3.17  14.5     0     1     5     4
#> 10  21.0     6 145     175  3.62  2.77  15.5     0     1     5     6
#> 11  15.8     8 301     335  3.54  3.57  14.6     0     1     5     8
#> 12  30.0     4 121     109  4.11  2.78  18.6     1     1     4     2